requirements.txt
streamlit
PyPDF2
openai
numpy
python-dotenv
